+++ src/Contracts/RedsysPayable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Creagia\LaravelRedsys\Contracts;
+
+interface RedsysPayable
+{
+    public function getTotalAmount(): int;
+
+    public function paidWithRedsys(): void;
+}